requirements.txt
streamlit
streamlit-mic-recorder
InstructorEmbedding
ctransformers
langchain
pypdfium2
librosa
pyyaml
accelerate
transformers
audio_recorder_streamlit
dotenv